Add Travis CI integration.
This commit introduces nascent Travis CI integration for Beefcake,
which will make the triage of antique pull requests far easier and
more reliable.

Due to how antique some of the code and the tooling around it is, I
have opted to use a direct makefile (.travis.makefile) to drive the
tests instead of using Travis with its implicit environmental
assumption.